More static casts

+++ mujoco_ros2_control/src/mujoco_system_interface.cpp
@@ -1000,7 +1000,7 @@ hardware_interface::return_type MujocoSystemInterface::write(const rclcpp::Time&
   // Clear plugin data, then let each plugin update as needed, in order. This enables plugins to read and
   // rewrite control inputs immediately before they are sent to the simulation. Namely, we have to zero
   // out xfrc_applied so plugins can update as needed.
-  mju_zero(mj_data_control_->xfrc_applied, 6 * simulation_->model()->nbody);
+  mju_zero(mj_data_control_->xfrc_applied, 6 * static_cast<int>(simulation_->model()->nbody));
   for (auto& plugin : plugin_instances_)
   {
     plugin->update(simulation_->model(), mj_data_control_);
